Fixing integration tests (2/2)
This fixes the remaining integration test failures following #157. The user creation fails to delete a user. This is because DD doesn't delete users; it disables them. When the test account is created,
it's already been created and disabled. To workaround the lack of deletion support (and to avoid creating a user every run), the test explicitly enable the user after creation. The other failure was for
the signature of the payload in the PagerDuty integration, which has been fixed. This commit should cut a new release, since it looks like this has been a bug for a while.

Closes #74

* User create/delete tests now enable user after creation
* PagerDuty integration updated