Make API key validation succeed without app key (#328)
Do not send the DD-APPLICATION-KEY header when the appkey is not set (ie. set to an empty string) on the client.

Sending a GET request to the api/v1/validate endpoint with a valid DD-API-KEY and no DD-APPLICATION-KEY succeeds.
Sending a GET request to the api/v1/validate endpoint with a valid DD-API-KEY and DD-APPLICATION-KEY set to an empty string fails.

This PR allows the API key validation method to succeed even when only an API key is used (which is enough in some use cases where an app key is not needed to access relevant endpoints).